// src/ts/types.ts
export type Velocity = {
  x: number
  y: number
}

export type Point = {
  x: number
  y: number
  radius: number
}

export type Stats = {
  score: number
  projectiles: number
  eliminations: number
  deaths: number
  time: number
}

export type Dimensions = {
  width: number
  height: number
}

export type Coords = {
  x: number
  y: number
}